+++ crates/tauri-plugin-pytauri/guest-js/src/index.ts
@@ -1,4 +1,4 @@
-import { invoke } from "@tauri-apps/api/core";
+import { invoke, Channel as TauriChannel } from "@tauri-apps/api/core";
 
 const PY_INVOKE_TAURI_CMD = "plugin:pytauri|pyfunc";
 const PY_INVOKE_HEADER = "pyfunc";
@@ -10,6 +10,13 @@ const textDecoder = new TextDecoder();
 type RawHandlerBodyType = ArrayBuffer | Uint8Array;
 type RawHandlerReturnType = ArrayBuffer;
 
+/**
+ * Invokes a Python function through the Tauri IPC mechanism.
+ *
+ * @param funcName - The name of the Python function to invoke.
+ * @param body - The body to send to the Python function.
+ * @returns A promise resolving or rejecting to the backend response.
+ */
 export async function rawPyInvoke(
     funcName: string,
     body: RawHandlerBodyType
@@ -34,6 +41,17 @@ it's a bug for pytauri, please report this issue."
     }
 }
 
+/**
+ * Invokes a Python function through the Tauri IPC mechanism.
+ *
+ * This is wrapper around `rawPyInvoke` that handles JSON serialization and deserialization.
+ *
+ * @template T - The expected return type of the Python function.
+ * @param funcName - The name of the Python function to invoke.
+ * @param body - The body to send to the Python function. It will be JSON serialized.
+ * @returns A promise resolving or rejecting to the backend response. It will be JSON deserialized.
+ * If you dont want JSON deserialization, use `rawPyInvoke` instead.
+ */
 export async function pyInvoke<T>(
     funcName: string,
     body: object | RawHandlerBodyType
@@ -51,4 +69,32 @@ export async function pyInvoke<T>(
 
     const respJson = textDecoder.decode(resp);
     return JSON.parse(respJson);
+}
+
+type RawChannelMsg = ArrayBuffer;
+
+/**
+ * This class is a subclass of {@link TauriChannel}.
+ * For the {@link TauriChannel} used by `pytauri`, it always transmits {@link ArrayBuffer}.
+ * Therefore, this class adds the {@link addJsonListener} method to help deserialize messages.
+ *
+ * If you dont need that, you can use {@link TauriChannel} directly.
+ *
+ * @template T - The expected return type from Python.
+ */
+export class Channel<T = unknown> extends TauriChannel<RawChannelMsg> {
+    constructor() {
+        super();
+    }
+
+    /**
+     * Equivalent to {@link TauriChannel.onmessage}, but it JSON deserializes the message as object.
+     */
+    addJsonListener(handler: (response: T) => void) {
+        this.onmessage = (bytes) => {
+            const msgJson = textDecoder.decode(bytes);
+            const response: T = JSON.parse(msgJson);
+            handler(response);
+        };
+    }
 }