+++ scripts/python-verifytypes.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
-pnpm pyright --verifytypes tauri_app --ignoreexternal && \
 pnpm pyright --verifytypes codelldb --ignoreexternal && \
 pnpm pyright --verifytypes pyfuture --ignoreexternal && \
 pnpm pyright --verifytypes pyo3_utils --ignoreexternal && \