+++ crates/tauri-plugin-pytauri/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Update bound `tauri-plugin-pytauri-api` to `v0.2.0`. See it's `CHANGELOG.md` for more details.
+
 ## [0.1.0-beta.0]
 
 [unreleased]: https://github.com/WSH032/pytauri/tree/HEAD