+++ crates/tauri-plugin-pytauri/guest-js/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+### Added
+
+- [#50](https://github.com/WSH032/pytauri/pull/50) - feat: add `class Channel extends TauriChannel<ArrayBuffer>` for pytauri [channels ipc](https://tauri.app/develop/calling-frontend/#channels).
+
+### Docs
+
+- [#50](https://github.com/WSH032/pytauri/pull/50) - add tsdoc for all classes and functions.
+
 ## [0.1.0-beta.0]
 
 [unreleased]: https://github.com/WSH032/pytauri/tree/HEAD